Update MemoryPool to use a ulong instead of uint to support higher memory allocation

+++ src/Titan.ECS/App/AppCreationArgs.cs
@@ -2,10 +2,10 @@ namespace Titan.ECS.App;
 
 public record struct AppCreationArgs
 {
-    public const uint DefaultUnmanagedMemory = 1 * 1024 * 1024 * 1024;      // 1Gb
+    public const ulong DefaultUnmanagedMemory = 1UL * 1024 * 1024 * 1024;      // 1Gb
     public const uint DefaultResourcesMemory = 32 * 1024 * 1024;      // 32Mb
     public const uint DefaultMaxResourceTypes = 1_000;
-    public uint UnmanagedMemory { get; init; }
+    public ulong UnmanagedMemory { get; init; }
     public uint ResourcesMemory { get; init; }
     public uint MaxResourceTypes { get; init; }
